Revert "Move utilities to new utility.h"
This reverts commit 6979437be3bc2127a937a3627918e0b3f2d669b8.
